% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_plot}
\alias{build_plot}
\title{Generates Tasks and Questionnaire plots}
\usage{
build_plot(plot)
}
\arguments{
\item{plot}{character. Which data should be generated. "AUDIT", "DASS" or "SDL".}
}
\value{
A plot for RShiny implementation.
}
\description{
Generates ggplot with Tasks and Questionnaire scores of the general population
and the users own score.
}
\keyword{internal}
