% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_questionnaire.R
\name{build_questionnaire}
\alias{build_questionnaire}
\title{Generates a Questionnaire for Rshiny}
\usage{
build_questionnaire(i, quest)
}
\arguments{
\item{i}{variable that will be looped through in a for loop, or within lapply() function.}

\item{quest}{character. Which Questionnaire the function should generate.}
}
\value{
A Questionnaire with Radio buttons for RShiny implementation.
}
\description{
Loads the Questionnaires from the data/ package folder.
}
\details{
Use only within RShiny. See an example of shiny implementation below.
}
\note{
Only to be used within RShiny, specifically within a lapply function.
}
\examples{
ui <- fluidPage(
        lapply(1:8, function(i) {
          build_questionnaire(i, "AUDIT_part1")
          }),
        lapply(9:10, function(i) {
          build_questionnaire(i, "AUDIT_part2")
          })
       )


}
\keyword{internal}
