% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_regression_plot}
\alias{build_regression_plot}
\title{Generates an Regression plot}
\usage{
build_regression_plot(x, y)
}
\value{
A plot for RShiny implementation.
}
\description{
Generates a regression plotly with a chosen Questionnaire scores and chosen cognitive
task score of the general population and the users own score.
}
\keyword{internal}
