% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievedata.R
\name{merge_data_files}
\alias{merge_data_files}
\title{Merges Your Score into General Population Data}
\usage{
merge_data_files(filename)
}
\arguments{
\item{filename}{The file name of your data file that you want merged.
Make sure you use a character string}

\item{datawd}{The absolute filepath of your data folder containing general population data file.
Make sure you use a character string (e.g., "C:/...")}
}
\value{
A data frame of all merged and pre-processed data.
}
\description{
Takes your raw data .csv files and merges them into
one general population data file  containing all the important measures
}
\details{
Use on data generated by ProjectExplicit experiment.
}
\note{
Takes all .csv files from the data folder, even if some files do not contain all data
(e.g., participant did not finish the entire task). Only works on .csv type files.
}
\examples{
data_folder <- getwd()
filename <- "audit_outcome.csv"
merge_data_files(data_folder, filename)


}
