% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectExplicit.R
\docType{package}
\name{ProjectExplicit}
\alias{ProjectExplicit}
\title{The package Project Explicit has 2 functions}
\section{Merge files}{


\emph{merge_data_files()}  Merges and pre-processes the raw data files.
}

\section{Save Data}{


\emph{save_data()}  Saves the previously merged data frame into .csv file.
}

