% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievedata.R
\name{save_data}
\alias{save_data}
\title{Saves the previously merged data file}
\usage{
save_data(wd, filename = "general_data.csv")
}
\arguments{
\item{wd}{The absolute filepath of the folder in which you want the file to be saved.
Make sure you use a character string (e.g., "C:/...")}

\item{file_name}{Default: "stroop_data". The name of the saved file in character string}
}
\value{
A .csv file saved in specified directory.
}
\description{
Saves the pre-processed and merged Stroop deadline .csv data file into
a specified directory, under a chosen file name.
}
\details{
Use on data generated by JsPsych Stroop Deadline experiment.
}
\note{
Requires your dataframe to already be in your Environment.
You can run merge_data_files() if you do not have your dataframe ready.
}
\examples{
data_folder <- getwd()
merge_data_files(data_folder)
save_data(data_folder, "my_data")



}
