% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_App}
\alias{start_App}
\title{Run the ProjectExplicit App}
\usage{
start_App()
}
\value{
Starts the app
}
\description{
Starts the App.
}
\details{
For best experience open in browser,
although the app works in RStudio pop-up window too.
}
