% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plots.R
\name{build_plot}
\alias{build_plot}
\title{Generates an Alcohol Use plot}
\usage{
build_plot(plot)
}
\value{
A plot for RShiny implementation.
}
\description{
Generates ggplot with AUDIT Questionnaire scores of the general population
and the users own score.
}
\keyword{internal}
