% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_questionnaire.R
\name{build_questionnaire}
\alias{build_questionnaire}
\title{Generates a Questionnaire for Rshiny}
\usage{
build_questionnaire(i, ans)
}
\arguments{
\item{i}{The absolute filepath of the folder in which you want the file to be saved.
Make sure you use a character string (e.g., "C:/...")}
}
\value{
A Questionnaire with Radio buttons in shiny
}
\description{
Saves the pre-processed and merged Stroop deadline .csv data file into
a specified directory, under a chosen file name.
}
\details{
Use on data generated by JsPsych Stroop Deadline experiment.
}
\note{
Requires your dataframe to already be in your Environment.
You can run merge_data_files() if you do not have your dataframe ready.
}
\examples{
data_folder <- getwd()
merge_data_files(data_folder)
save_data(data_folder, "my_data")



}
