% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrievedata.R
\name{merge_data_files}
\alias{merge_data_files}
\title{Merges separate data files into one}
\usage{
merge_data_files(datawd)
}
\arguments{
\item{datawd}{The absolute filepath of your data folder containing raw participant data files.
Make sure you use a character string (e.g., "C:/...")}
}
\value{
A data frame of all merged and pre-processed raw Stroop Task data.
Specifically, practice trials accuracy, test trials accuracy,
proportion of no responses, and adaptive response window, for each subject.
}
\description{
Takes all separate participant raw data .csv files and merges them into
one pre-processed file  containing the important Stroop deadline task measures
(i.e., accuracy and response time)
}
\details{
Use on data generated by JsPsych Stroop Deadline experiment.
}
\note{
Takes all .csv files from the data folder, even if some files do not contain all data
(e.g., participant did not finish the entire task). Only works on .csv type files.
}
\examples{
data_folder <- getwd()
merge_data_files(data_folder)


}
